style: Reformat source

+++ src/main/java/org/yliadevelopment/network/BinaryHelper.java
@@ -1,8 +1,9 @@
 package org.yliadevelopment.network;
 
 public final class BinaryHelper {
-    
-    private BinaryHelper() {}
+
+    private BinaryHelper() {
+    }
 
     public static boolean readBoolean(byte b) {
         return b == 1;
@@ -13,10 +14,8 @@ public final class BinaryHelper {
     }
 
     public static int readUnsignedShort(byte[] buf) {
-        return (int) (
-            ((buf[0] & 0xFF) << 8) +
-            ((buf[1] & 0xFF) << 0)
-        );
+        return ((buf[0] & 0xFF) << 8) +
+                ((buf[1] & 0xFF) << 0);
     }
 
     public static short readShort(byte[] buf) {
@@ -24,10 +23,8 @@ public final class BinaryHelper {
     }
 
     public static int readUnsignedLShort(byte[] buf) {
-        return (int) (
-            ((buf[1] & 0xFF) << 8) +
-            ((buf[0] & 0xFF) << 0)
-        );
+        return ((buf[1] & 0xFF) << 8) +
+                ((buf[0] & 0xFF) << 0);
     }
 
     public static short readLShort(byte[] buf) {
@@ -35,39 +32,35 @@ public final class BinaryHelper {
     }
 
     public static int readTriad(byte[] buf) {
-        return readInt(new byte[] {
-            0x00,
-            buf[0],
-            buf[1],
-            buf[2]
+        return readInt(new byte[]{
+                0x00,
+                buf[0],
+                buf[1],
+                buf[2]
         });
     }
 
     public static int readLTriad(byte[] buf) {
-        return readInt(new byte[] {
-            buf[0],
-            buf[1],
-            buf[2],
-            0x00
+        return readInt(new byte[]{
+                buf[0],
+                buf[1],
+                buf[2],
+                0x00
         });
     }
 
     public static int readInt(byte[] buf) {
-        return (int) (
-            ((buf[0] & 0xFF) << 24) +
-            ((buf[1] & 0xFF) << 16) +
-            ((buf[2] & 0xFF) <<  8) +
-            ((buf[3] & 0xFF) <<  0)
-        );
+        return ((buf[0] & 0xFF) << 24) +
+                ((buf[1] & 0xFF) << 16) +
+                ((buf[2] & 0xFF) << 8) +
+                ((buf[3] & 0xFF) << 0);
     }
 
     public static int readLInt(byte[] buf) {
-        return (int) (
-            ((buf[3] & 0xFF) << 24) + 
-            ((buf[2] & 0xFF) << 16) +
-            ((buf[1] & 0xFF) <<  8) +
-            ((buf[0] & 0xFF) <<  0)
-        );
+        return ((buf[3] & 0xFF) << 24) +
+                ((buf[2] & 0xFF) << 16) +
+                ((buf[1] & 0xFF) << 8) +
+                ((buf[0] & 0xFF) << 0);
     }
 
     public static float readFloat(byte[] buf) {
@@ -79,29 +72,25 @@ public final class BinaryHelper {
     }
 
     public static long readLong(byte[] buf) {
-        return (long) (
-            ((buf[0] & 0xFF) << 56) +
-            ((buf[1] & 0xFF) << 48) +
-            ((buf[2] & 0xFF) << 40) +
-            ((buf[3] & 0xFF) << 32) +
-            ((buf[4] & 0xFF) << 24) +
-            ((buf[5] & 0xFF) << 16) +
-            ((buf[6] & 0xFF) <<  8) +
-            ((buf[7] & 0xFF) <<  0)
-        );
+        return ((long) (buf[0] & 0xFF) << 56) +
+                ((long) (buf[1] & 0xFF) << 48) +
+                ((long) (buf[2] & 0xFF) << 40) +
+                ((long) (buf[3] & 0xFF) << 32) +
+                ((long) (buf[4] & 0xFF) << 24) +
+                ((buf[5] & 0xFF) << 16) +
+                ((buf[6] & 0xFF) << 8) +
+                ((buf[7] & 0xFF) << 0);
     }
 
     public static long readLLong(byte[] buf) {
-        return (long) (
-            ((buf[7] & 0xFF) << 56) +
-            ((buf[6] & 0xFF) << 48) +
-            ((buf[5] & 0xFF) << 40) +
-            ((buf[4] & 0xFF) << 32) +
-            ((buf[3] & 0xFF) << 24) +
-            ((buf[2] & 0xFF) << 16) +
-            ((buf[1] & 0xFF) <<  8) +
-            ((buf[0] & 0xFF) <<  0)
-        );
+        return ((long) (buf[7] & 0xFF) << 56) +
+                ((long) (buf[6] & 0xFF) << 48) +
+                ((long) (buf[5] & 0xFF) << 40) +
+                ((long) (buf[4] & 0xFF) << 32) +
+                ((long) (buf[3] & 0xFF) << 24) +
+                ((buf[2] & 0xFF) << 16) +
+                ((buf[1] & 0xFF) << 8) +
+                ((buf[0] & 0xFF) << 0);
     }
 
     public static double readDouble(byte[] buf) {
@@ -123,52 +112,52 @@ public final class BinaryHelper {
     public static byte[] writeShort(int s) {
         s &= 0xFFFF;
 
-        return new byte[] {
-            (byte) ((s >> 8) & 0xFF),
-            (byte) ((s >> 0) & 0xFF)
+        return new byte[]{
+                (byte) ((s >> 8) & 0xFF),
+                (byte) ((s >> 0) & 0xFF)
         };
     }
 
     public static byte[] writeLShort(int s) {
         s &= 0xFFFF;
-        
-        return new byte[] {
-            (byte) ((s >> 0) & 0xFF),
-            (byte) ((s >> 8) & 0xFF)
+
+        return new byte[]{
+                (byte) ((s >> 0) & 0xFF),
+                (byte) ((s >> 8) & 0xFF)
         };
     }
 
     public static byte[] writeTriad(int t) {
-        return new byte[] {
-            (byte) ((t >> 16) & 0xFF),
-            (byte) ((t >>  8) & 0xFF),
-            (byte) ((t >>  0) & 0xFF)
+        return new byte[]{
+                (byte) ((t >> 16) & 0xFF),
+                (byte) ((t >> 8) & 0xFF),
+                (byte) ((t >> 0) & 0xFF)
         };
     }
 
     public static byte[] writeLTriad(int t) {
-        return new byte[] {
-            (byte) ((t >>  0) & 0xFF),
-            (byte) ((t >>  8) & 0xFF),
-            (byte) ((t >> 16) & 0xFF)
+        return new byte[]{
+                (byte) ((t >> 0) & 0xFF),
+                (byte) ((t >> 8) & 0xFF),
+                (byte) ((t >> 16) & 0xFF)
         };
     }
 
     public static byte[] writeInt(int i) {
-        return new byte[] {
-            (byte) ((i >> 24) & 0xFF),
-            (byte) ((i >> 16) & 0xFF),
-            (byte) ((i >>  8) & 0xFF),
-            (byte) ((i >>  0) & 0xFF)
+        return new byte[]{
+                (byte) ((i >> 24) & 0xFF),
+                (byte) ((i >> 16) & 0xFF),
+                (byte) ((i >> 8) & 0xFF),
+                (byte) ((i >> 0) & 0xFF)
         };
     }
 
     public static byte[] writeLInt(int i) {
-        return new byte[] {
-            (byte) ((i >>  0) & 0xFF),
-            (byte) ((i >>  8) & 0xFF),
-            (byte) ((i >> 16) & 0xFF),
-            (byte) ((i >> 24) & 0xFF)
+        return new byte[]{
+                (byte) ((i >> 0) & 0xFF),
+                (byte) ((i >> 8) & 0xFF),
+                (byte) ((i >> 16) & 0xFF),
+                (byte) ((i >> 24) & 0xFF)
         };
     }
 
@@ -181,27 +170,27 @@ public final class BinaryHelper {
     }
 
     public static byte[] writeLong(long l) {
-        return new byte[] {
-            (byte) ((l >> 56) & 0xFF),
-            (byte) ((l >> 48) & 0xFF),
-            (byte) ((l >> 32) & 0xFF),
-            (byte) ((l >> 24) & 0xFF),
-            (byte) ((l >> 16) & 0xFF),
-            (byte) ((l >>  8) & 0xFF),
-            (byte) ((l >>  0) & 0xFF)
+        return new byte[]{
+                (byte) ((l >> 56) & 0xFF),
+                (byte) ((l >> 48) & 0xFF),
+                (byte) ((l >> 32) & 0xFF),
+                (byte) ((l >> 24) & 0xFF),
+                (byte) ((l >> 16) & 0xFF),
+                (byte) ((l >> 8) & 0xFF),
+                (byte) ((l >> 0) & 0xFF)
         };
     }
 
     public static byte[] writeLLong(long l) {
-        return new byte[] {
-            (byte) ((l >>  0) & 0xFF),
-            (byte) ((l >>  8) & 0xFF),
-            (byte) ((l >> 16) & 0xFF),
-            (byte) ((l >> 24) & 0xFF),
-            (byte) ((l >> 32) & 0xFF),
-            (byte) ((l >> 40) & 0xFF),
-            (byte) ((l >> 48) & 0xFF),
-            (byte) ((l >> 56) & 0xFF)
+        return new byte[]{
+                (byte) ((l >> 0) & 0xFF),
+                (byte) ((l >> 8) & 0xFF),
+                (byte) ((l >> 16) & 0xFF),
+                (byte) ((l >> 24) & 0xFF),
+                (byte) ((l >> 32) & 0xFF),
+                (byte) ((l >> 40) & 0xFF),
+                (byte) ((l >> 48) & 0xFF),
+                (byte) ((l >> 56) & 0xFF)
         };
     }
